changed temperature fields

+++ monitor/models.py
@@ -3,7 +3,7 @@ from django.contrib.auth.models import User
 
 
 class Raspberry(models.Model):
-    temperature = models.DecimalField(decimal_places=1, max_digits=3)
+    temperature = models.FloatField()
     memory_used = models.IntegerField()
 
     def __str__(self):
